+++ src/com/maddyhome/idea/vim/helper/EditorHelper.java
@@ -26,6 +26,7 @@ import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.testFramework.LightVirtualFile;
 import com.maddyhome.idea.vim.common.IndentConfig;
 import com.maddyhome.idea.vim.common.TextRange;
+import com.maddyhome.idea.vim.ui.ExEntryPanel;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -670,9 +671,17 @@ public class EditorHelper {
   public static boolean scrollVisualLineToBottomOfScreen(@NotNull Editor editor, int visualLine) {
     final ScrollingModel scrollingModel = editor.getScrollingModel();
     int inlayHeight = getHeightOfVisualLineInlays(editor, visualLine, false);
+    int exPanelHeight = 0;
+    int exPanelWithoutShortcutsHeight = 0;
+    if (ExEntryPanel.getInstance().isActive()) {
+      exPanelHeight = ExEntryPanel.getInstance().getHeight();
+    }
+    if (ExEntryPanel.getInstanceWithoutShortcuts().isActive()) {
+      exPanelWithoutShortcutsHeight = ExEntryPanel.getInstanceWithoutShortcuts().getHeight();
+    }
     int y = editor.visualLineToY(visualLine);
     int verticalPos = scrollingModel.getVerticalScrollOffset();
-    int height = inlayHeight + editor.getLineHeight();
+    int height = inlayHeight + editor.getLineHeight() + exPanelHeight + exPanelWithoutShortcutsHeight;
 
     Rectangle visibleArea = scrollingModel.getVisibleArea();