+++ CHANGES.md
@@ -39,6 +39,7 @@ usual beta standards.
 * [VIM-1855](https://youtrack.jetbrains.com/issue/VIM-1855) Fix initialization error
 * [VIM-1853](https://youtrack.jetbrains.com/issue/VIM-1853) Fix marks for disposed projects
 * [VIM-1858](https://youtrack.jetbrains.com/issue/VIM-1858) Fix imap for autocomplete
+* [VIM-1362](https://youtrack.jetbrains.com/issue/VIM-1362) Search with confirm doesn't scroll down far enough
 
 
 0.54, 2019-11-20